+++ src/webapp/src/api/device.ts
@@ -1,16 +1,13 @@
 import api from "../utils/api";
 import {isAuthenticated} from "./authenticate";
-import {getAuthenticationHeader} from "../utils/auth";
 import {AxiosError} from "axios";
+import {Device, DeviceCoordinate} from "../utils/dtos";
 
-export async function getDevices() {
+export async function getDevices(): Promise<Device[]> {
     if (!(await isAuthenticated()))
         throw new Error("Nu eşti autentificat.");
     try {
-        const response = await api.get(
-            "/admin/device/all",
-            {headers: getAuthenticationHeader()}
-        );
+        const response = await api.get("/admin/device/all");
         return response.data;
     } catch (error) {
         throw new Error("Eroare în cadrul serverului. Vă rugăm să încercați mai târziu.");
@@ -23,8 +20,7 @@ export async function createDevice(name: string, location: string) {
     try {
         const response = await api.post(
             "/admin/device",
-            {name, location},
-            {headers: getAuthenticationHeader()}
+            {name, location}
         );
         return response.data;
     } catch (error) {
@@ -38,10 +34,7 @@ export async function deleteDevice(deviceId: number) {
     if (!(await isAuthenticated()))
         throw new Error("Nu eşti autentificat.");
     try {
-        const response = await api.delete(
-            `/admin/device/${deviceId}`,
-            {headers: getAuthenticationHeader()}
-        );
+        const response = await api.delete(`/admin/device/${deviceId}`);
     } catch (error) {
         throw new Error("Eroare în cadrul serverului. Vă rugăm să încercați mai târziu.");
     }
@@ -53,12 +46,15 @@ export async function updateDevice(deviceId: number, name: string, location: str
     try {
         const response = await api.put(
             `/admin/device/${deviceId}`,
-            {name, location},
-            {headers: getAuthenticationHeader()}
+            {name, location}
         );
     } catch (error) {
         if (error instanceof AxiosError && error.response?.status === 400)
             throw new Error("Numele dispozitivului există deja.");
         throw new Error("Eroare în cadrul serverului. Vă rugăm să încercați mai târziu.");
     }
+}
+
+export async function getDeviceCoordinates(): Promise<DeviceCoordinate[]> {
+  return (await api.get("/public/device/coordinate")).data;
 }
